Tasks/Notes Box
From Steve:

I have just noticed though that when I go into the task, it shows me a list of all the notes created and has the radio button checked where we have set the reminder.

Remove the words "Add" from each of the tabs.

The note which has been added using "set reminder" in the notes section makes that so it is statically checked and can only be change in the edit reminder of the notes section

I think it is a good idea that other notes are also available to select but they need to be multi select radio buttons (or multi check boxes) that can be turned on and off so it gives the user the option to indicate other notes attached to this task.

I hope this makes sense